Deduplicate script loader test helpers
