+++ tests/unit/test_verify_modules_signature_script.py
@@ -1,10 +1,10 @@
 from __future__ import annotations
 
-import builtins
-import importlib.util
 from pathlib import Path
 
 import yaml
+
+from tests.unit._script_test_utils import block_contract_imports, load_module_from_path
 
 
 REPO_ROOT = Path(__file__).resolve().parents[2]
@@ -12,23 +12,11 @@ SCRIPT_PATH = REPO_ROOT / "scripts" / "verify-modules-signature.py"
 
 
 def _load_verify_script():
-    spec = importlib.util.spec_from_file_location("verify_modules_signature", SCRIPT_PATH)
-    if spec is None or spec.loader is None:
-        raise AssertionError("Unable to load verify-modules-signature.py")
-    module = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(module)
-    return module
+    return load_module_from_path("verify_modules_signature", SCRIPT_PATH)
 
 
 def test_verify_script_loads_without_beartype_or_icontract(monkeypatch) -> None:
-    original_import = builtins.__import__
-
-    def raising_import(name, globalns=None, localns=None, fromlist=(), level=0):
-        if name in {"beartype", "icontract"}:
-            raise ImportError(f"blocked import for {name}")
-        return original_import(name, globalns, localns, fromlist, level)
-
-    monkeypatch.setattr(builtins, "__import__", raising_import)
+    block_contract_imports(monkeypatch)
 
     verify_script = _load_verify_script()
